modulith(phase-5): extract team module
- Move Team entity, TeamRepository, TeamRegistrationService, TeamController, DTOs to team package
- Decouple Team JPA: @ManyToOne Quiz -> @Column Long quizId, remove submissions collection
- Create TeamFacade, TeamInfoDto, AccessCode value object
- Create TeamRegisteredEvent, TeamRemovedEvent, TeamScoreResetEvent
- Update TeamRepository: findByQuiz(Quiz) -> findByQuizId(Long)
- Update TeamRegistrationService to use QuizFacade instead of QuizRepository
- Fix Question.isCorrectAnswer() to trim correctKey (bug found by property test)
- Update all external references and test files
- 215 tests passing

+++ backend/src/test/java/com/intelliquiz/api/application/services/EntityValidationPropertyTest.java
@@ -3,6 +3,7 @@ package com.intelliquiz.api.application.services;
 import com.intelliquiz.api.domain.entities.*;
 import com.intelliquiz.api.quiz.internal.domain.entities.Quiz;
 import com.intelliquiz.api.quiz.internal.domain.entities.Question;
+import com.intelliquiz.api.team.internal.domain.entities.Team;
 import com.intelliquiz.api.user.internal.domain.entities.User;
 import com.intelliquiz.api.shared.enums.*;
 import net.jqwik.api.*;
@@ -78,7 +79,7 @@ public class EntityValidationPropertyTest {
      */
     @Property(tries = 20)
     void teamScoreCannotBeNegative(@ForAll("negativeIntegers") int negativeScore) {
-        Team team = new Team(null, "Team", "ABC-123");
+        Team team = new Team((Long) null, "Team", "ABC-123");
         team.setTotalScore(negativeScore);
         
         assertThatThrownBy(team::validateScore)
@@ -91,7 +92,7 @@ public class EntityValidationPropertyTest {
      */
     @Property(tries = 20)
     void teamScoreValidationAcceptsNonNegative(@ForAll("nonNegativeIntegers") int score) {
-        Team team = new Team(null, "Team", "ABC-123");
+        Team team = new Team((Long) null, "Team", "ABC-123");
         team.setTotalScore(score);
         
         // Should not throw
